Fix WebView2 wallpaper scaling using reflection..
.. finally fixed it 😅. Previous implementation was failing with various display arrangement/primary screen combination.

Used reflection since RasterizationScale is not exposed.

+++ src/Lively/Lively.Player.WebView2/StartArgs.cs
@@ -34,7 +34,7 @@ namespace Lively.Player.WebView2
         [Option("wallpaper-scale",
         Required = false,
         HelpText = "Wallpaper scale factor.")]
-        public int? Scale { get; set; }
+        public double? Scale { get; set; }
 
         [Option("wallpaper-audio",
         Default = false,
